Add project api (#13)
* Add React Query and update project fetching in dashboard

- Integrate @tanstack/react-query for data fetching
- Add new provider component to wrap QueryClientProvider and ThemeProvider
- Implement useGetProjects hook for fetching dashboard projects
- Update dashboard page to use dynamic project data
- Add new project API route with filtering and pagination
- Remove hardcoded project data
- Add @radix-ui/react-label to dependencies

* Add project creation functionality and UI improvements

- Implement CreateProject component with dynamic placeholder text
- Add loading state to Button component
- Update Content and HoverEffect components to handle undefined data
- Create new user API route for fetching authenticated user
- Refactor project-related types and components
- Add sarcastic placeholder generation for project creation

* Add UI components and project hooks for project management

- Create Label and Textarea UI components using Radix UI and class-variance-authority
- Implement useProject hook with create and get project mutations/queries
- Add project type definitions for consistent data handling

* Refactor user API route and extract getUser helper function

- Move getUser function to a separate helper file
- Update import statements in project route to use the new helper location
- Simplify user route by importing the extracted helper function
