+++ app/api/user/route.ts
@@ -1,29 +1,5 @@
-import { auth } from "@clerk/nextjs/server";
 import { NextResponse } from "next/server";
-import { prisma } from "@/lib/prismaClient";
-
-export const getUser = async () => {
-  const { userId: clerkId } = await auth();
-  if (!clerkId) {
-    throw new Error("Unauthorized");
-  }
-
-  const user = await prisma.user.findUnique({
-    where: {
-      externalId: clerkId,
-    },
-    select: {
-      id: true,
-    },
-  });
-
-  if (!user) {
-    throw new Error("User not found");
-  }
-
-  return user;
-};
-
+import { getUser } from "./helper";
 export async function GET() {
   try {
     const user = await getUser();
